refactor : 1차 멘토님께 조언받은 부분 수정
