+++ src/components/header/header.tsx
@@ -1,13 +1,10 @@
-import { useLocation, useNavigate } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import "./header.css";
 import ImgHomeLogo from "../../assets/logo/img_logo.svg";
+import NavItem from "../navitem.tsx";
 
 function Header() {
-  const location = useLocation();
   const navigate = useNavigate();
-  const isCommunityPage = location.pathname === "/board";
-  const isMarketPage =
-    location.pathname === "/items" || location.pathname === "/additem"; // 수정된 부분
 
   const handleLoginButtonClick = () => {
     navigate("/login");
@@ -21,22 +18,8 @@ function Header() {
         </a>
         <nav className="header-nav">
           <ul className="header-nav-container">
-            <li className="header-nav-item">
-              <a
-                href="/board"
-                className={`nav-link ${isCommunityPage ? "active" : ""}`}
-              >
-                자유게시판
-              </a>
-            </li>
-            <li className="header-nav-item">
-              <a
-                href="/items"
-                className={`nav-link ${isMarketPage ? "active" : ""}`}
-              >
-                중고마켓
-              </a>
-            </li>
+            <NavItem to="/board">자유게시판</NavItem>
+            <NavItem to="/items">중고마켓</NavItem>
           </ul>
         </nav>
         <button className="header-login-btn" onClick={handleLoginButtonClick}>
