+++ src/pages/login/login.css
@@ -66,7 +66,7 @@
 }
 
 .password-eye {
-  background-image: url("../../assets/logo/ic-noeye.png");
+  background-image: url("../../assets/logo/ic-noeye.svg");
   background-position: center;
   background-size: contain;
   background-repeat: no-repeat;
